review: make call_contract_function private
